+++ test/runtests.jl
@@ -8,5 +8,6 @@ using Test
     include("testsets/Errors.jl")
     include("testsets/Responses.jl")
     include("testsets/Router.jl")
+    include("testsets/Server.jl")
     include("testsets/Extensions.jl")
 end